Danny's first commit
Change namespaces of Item and Inventory => Items, Inventories
Inventory play around
