--- Spoops/Game/Inventory/Inventory.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Spoops.Game.Inventory
-{
-    public abstract class Inventory
-    {
-
-    }
-}
