+++ Spoops/Spoops.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spoops", "Spoops.csproj", "{71DBD394-86C6-4A16-B92C-DB41E6B6E5E9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{71DBD394-86C6-4A16-B92C-DB41E6B6E5E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71DBD394-86C6-4A16-B92C-DB41E6B6E5E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71DBD394-86C6-4A16-B92C-DB41E6B6E5E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71DBD394-86C6-4A16-B92C-DB41E6B6E5E9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
